TIP-221: Move AttributeRequirementFactory
