+++ spec/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/Family/SetAttributeRequirementsSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Family;
 use Akeneo\Component\Batch\Model\JobExecution;
 use Akeneo\Component\Batch\Model\StepExecution;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Factory\AttributeRequirementFactory;
+use Pim\Component\Catalog\Factory\AttributeRequirementFactory;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\AttributeRequirementInterface;
 use Pim\Component\Catalog\Model\ChannelInterface;
