+++ src/Pim/Bundle/EnrichBundle/Connector/Processor/MassEdit/Family/SetAttributeRequirements.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\EnrichBundle\Connector\Processor\MassEdit\Family;
 
 use InvalidArgumentException;
-use Pim\Bundle\CatalogBundle\Factory\AttributeRequirementFactory;
+use Pim\Component\Catalog\Factory\AttributeRequirementFactory;
 use Pim\Bundle\EnrichBundle\Connector\Processor\AbstractProcessor;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
 use Pim\Component\Catalog\Repository\ChannelRepositoryInterface;
